test(web-integration): split aiAct plan cache fallback AI test
