+++ packages/web-integration/tests/ai/web/puppeteer/cache-functionality.test.ts
@@ -5,7 +5,6 @@ import { sleep } from '@midscene/core/utils';
 import { uuid } from '@midscene/shared/utils';
 import yaml from 'js-yaml';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { getFixturePath } from './test-utils';
 import { launchPage } from './utils';
 
 vi.setConfig({
@@ -419,58 +418,6 @@ describe('Cache Operation Tests', () => {
 
     expect(secondLoadPlanningSpy).toHaveBeenCalled();
     expect(secondActionSpy).not.toHaveBeenCalled();
-  });
-
-  it('should fall back when a cached optional popup plan no longer matches the page', async () => {
-    const prompt = 'click the Complete purchase button';
-    const staleWorkflow = `tasks:
-  - name: click the Complete purchase button
-    flow:
-      - aiTap: the close button in the optional promotion popup
-      - aiTap: the Complete purchase button
-`;
-    const cacheId = `optional-popup-plan-cache-${uuid()}`;
-    const staleCache = new TaskCache(cacheId, true);
-    cacheFilePathToCleanup = staleCache.cacheFilePath;
-    staleCache.appendCache({
-      type: 'plan',
-      prompt,
-      yamlWorkflow: staleWorkflow,
-    });
-
-    const { originPage, reset } = await launchPage(
-      `file://${getFixturePath('optional-popup-cache.html')}`,
-    );
-    resetFn = reset;
-
-    agent = new PuppeteerAgent(originPage, {
-      cache: { id: cacheId },
-    });
-    const runYamlSpy = vi.spyOn(agent, 'runYaml');
-    const actionSpy = vi.spyOn((agent as any).taskExecutor, 'action');
-
-    await agent.aiAct(prompt);
-
-    expect(runYamlSpy).toHaveBeenCalledOnce();
-    expect(actionSpy).toHaveBeenCalled();
-    await expect(
-      originPage.$eval('#status', (element) => element.textContent),
-    ).resolves.toBe('Order completed');
-
-    const cacheContent = yaml.load(
-      fs.readFileSync(cacheFilePathToCleanup!, 'utf-8'),
-    ) as {
-      caches: Array<{
-        type: string;
-        prompt: string;
-        yamlWorkflow?: string;
-      }>;
-    };
-    const planCaches = cacheContent.caches.filter(
-      (cache) => cache.type === 'plan' && cache.prompt === prompt,
-    );
-    expect(planCaches).toHaveLength(1);
-    expect(planCaches[0].yamlWorkflow).not.toBe(staleWorkflow);
   });
 
   it('should handle cache with cacheable: false option', async () => {
